Refactor MCP server
